Move BlockId into codechain-types
Coordinator and modules need BlockId, so we should move it to
codechain-types crate in order to avoid cyclic dependecy.
