+++ core/src/client/importer.rs
@@ -22,13 +22,12 @@ use crate::consensus::CodeChainEngine;
 use crate::error::Error;
 use crate::miner::{Miner, MinerService};
 use crate::service::ClientIoMessage;
-use crate::types::BlockId;
 use crate::verification::queue::{BlockQueue, HeaderQueue};
 use crate::verification::{PreverifiedBlock, Verifier};
 use crate::views::{BlockView, HeaderView};
 use cio::IoChannel;
 use ctypes::header::{Header, Seal};
-use ctypes::BlockHash;
+use ctypes::{BlockHash, BlockId};
 use kvdb::DBTransaction;
 use parking_lot::{Mutex, MutexGuard};
 use rlp::Encodable;
