+++ core/src/consensus/tendermint/worker.rs
@@ -37,13 +37,12 @@ use crate::snapshot_notify::NotifySender as SnapshotNotifySender;
 use crate::transaction::UnverifiedTransaction;
 use crate::types::BlockStatus;
 use crate::views::BlockView;
-use crate::BlockId;
 use ckey::{verify, Ed25519Public as Public, Signature};
 use cnetwork::{EventSender, NodeId};
 use crossbeam_channel as crossbeam;
 use ctypes::transaction::Transaction;
 use ctypes::util::unexpected::Mismatch;
-use ctypes::{BlockHash, BlockNumber, Header};
+use ctypes::{BlockHash, BlockId, BlockNumber, Header};
 use primitives::Bytes;
 use rlp::{Encodable, Rlp};
 use std::cell::Cell;
