+++ core/src/miner/miner.rs
@@ -26,12 +26,12 @@ use crate::consensus::{CodeChainEngine, EngineType};
 use crate::error::Error;
 use crate::scheme::Scheme;
 use crate::transaction::{PendingVerifiedTransactions, UnverifiedTransaction, VerifiedTransaction};
-use crate::types::{BlockId, TransactionId};
+use crate::types::TransactionId;
 use ckey::{Ed25519Private as Private, Ed25519Public as Public, Password, PlatformAddress};
 use cstate::{FindDoubleVoteHandler, TopLevelState, TopStateView};
 use ctypes::errors::HistoryError;
 use ctypes::transaction::{IncompleteTransaction, Transaction};
-use ctypes::{BlockHash, TxHash};
+use ctypes::{BlockHash, BlockId, TxHash};
 use kvdb::KeyValueDB;
 use parking_lot::{Mutex, RwLock};
 use primitives::Bytes;
