+++ core/src/service.rs
@@ -18,10 +18,9 @@ use crate::client::{Client, ClientConfig};
 use crate::error::Error;
 use crate::miner::Miner;
 use crate::scheme::Scheme;
-use crate::BlockId;
 use cio::{IoContext, IoHandler, IoHandlerResult, IoService};
 use ctimer::TimerApi;
-use ctypes::BlockHash;
+use ctypes::{BlockHash, BlockId};
 use kvdb::KeyValueDB;
 use primitives::Bytes;
 use std::sync::Arc;
