+++ informer/Cargo.toml
@@ -16,6 +16,7 @@ parking_lot = "0.6.0"
 rlp = { git = "https://github.com/CodeChain-io/rlp.git", version = "0.4" }
 cinfo_courier = { package = "codechain_informer_courier", path = "../informer_courier" }
 ccore = { package = "codechain-core", path = "../core" }
+ctypes = { package = "codechain-types", path = "../types" }
 crpc = { package = "codechain-rpc", path = "../rpc" }
 tokio = { version = "0.2.13", features = ["full"] }
 serde = "1.0"
